Revert "move retry logic out into lib"
This reverts commit 4adc6775f0e3f08f20a806cbaf86e03ba65d98ed.
