+++ lib/errors.js
@@ -35,13 +35,6 @@ class BlobSHAMismatchError extends Error {
   }
 }
 
-class MaxRetryError extends Error {
-  constructor(message, retries) {
-    super(message)
-    this.retries = retries
-  }
-}
-
 export {
-  OutOfBoundsError, ElasticSearchError, NullMetadataError, BlobSHAMismatchError, MaxRetryError,
+  OutOfBoundsError, ElasticSearchError, NullMetadataError, BlobSHAMismatchError,
 }
