+++ lib/result.js
@@ -1,5 +1,4 @@
 import * as Errors from './errors.js'
-import Retry from './retry.js'
 
 class Result {
   constructor(input) {
@@ -12,34 +11,27 @@ class Result {
     }
     this.metadataURL = `https://archive.org/metadata/${this.identifier}`
     this.files = []
-    this.retry = 5
   }
 
   async metadata() {
-    const cb = async () => {
-      const response = await fetch(this.metadataURL)
+    const response = await fetch(this.metadataURL)
 
-      if (!response.ok) {
-        throw new Error(`an error occured fetching metadata: ${response}`)
-      }
-
-      const data = await response.json()
-
-      if (Object.keys(data).length === 0) {
-        throw new Errors.NullMetadataError(`no metadata exists for ${this.identifier}`)
-      }
-
-      if (typeof data.metadata.subject === 'string') {
-        data.metadata.subject = data.metadata.subject.split(';')
-      }
-
-      return data
+    if (!response.ok) {
+      throw new Error(`an error occured fetching metadata: ${response}`)
     }
-    const data = await Retry(0, this.retry, cb, [Errors.NullMetadataError])
 
+    const data = await response.json()
+
+    if (Object.keys(data).length === 0) {
+      throw new Errors.NullMetadataError(`no metadata exists for ${this.identifier}`)
+    }
+
+    if (typeof data.metadata.subject === 'string') {
+      data.metadata.subject = data.metadata.subject.split(';')
+    }
     Object.assign(this, data.metadata)
     this.files = data.files
-    this.downloadURL = `http://${data.server}${data.dir}`
+    this.downloadURL = `https://${data.server}${data.dir}`
   }
 
   * download(start = 0, end = Infinity) {
@@ -55,7 +47,7 @@ class Result {
 
     for (let i = start; i < len; i += 1) {
       count += 1
-      const cb = () => fetch(`${this.downloadURL}/${this.files[i].name}`).then((response) => {
+      yield fetch(`${this.downloadURL}/${this.files[i].name}`).then((response) => {
         if (!response.ok) {
           throw new Error(`error occured while downloading file ${response}`)
         }
@@ -72,13 +64,12 @@ class Result {
         }
         throw new Errors.BlobSHAMismatchError(`downloaded file ${this.files[i].name} SHA-1 does not match`, desired, current)
       })
-      yield Retry(0, this.retry, cb, [Errors.BlobSHAMismatchError])
     }
     return count
   }
 
   renderCallback(start = 0, end = Infinity, cb = (blob) => blob) {
-    this.metadata.then(() => {
+    this.metadata().then(() => {
       for (const blobProm of this.download(start, end)) {
         blobProm.then((blob) => cb(blob))
       }
