+++ lib/search.js
@@ -1,6 +1,5 @@
 import * as Errors from './errors.js'
 import Result from './result.js'
-import Retry from './retry.js'
 
 /* eslint-disable max-classes-per-file */
 class Search {
@@ -60,6 +59,31 @@ class Search {
     return data.response.docs
   }
 
+  searchRetry(page = 1, rows = 0, retry = 0) {
+    const delay = (count) => new Promise(
+      (resolve) => {
+        setTimeout(resolve, 10 ** count)
+      },
+    )
+    return this.search(page, rows).catch(
+      (reason) => {
+        if (reason instanceof Errors.ElasticSearchError ||
+          reason instanceof Errors.OutOfBoundsError ||
+          reason instanceof TypeError) {
+          return Promise.reject(reason)
+        }
+        return delay(retry).then(
+          () => {
+            if (retry > this.retry) {
+              return Promise.reject(reason)
+            }
+            return this.searchRetry(page, rows, retry + 1)
+          },
+        )
+      },
+    )
+  }
+
   // eslint-disable-next-line compat/compat
   [Symbol.asyncIterator]() {
     let { page } = this
@@ -67,11 +91,7 @@ class Search {
       next: async () => {
         const remainRows = this.end - (page - 1) * this.rows
         if (page * this.rows <= this.end || remainRows > 0) {
-          const result = await Retry(
-            0,
-            this.retry,
-            this.search,
-            [Errors.ElasticSearchError, Errors.OutOfBoundsError],
+          const result = await this.searchRetry(
             page,
             Math.min(remainRows, this.rows),
           )
