+++ lib/test.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 import Search from './search.js'
+import Result from './result.js'
 import Fields from './fields.js'
 import * as Query from './query.js'
 
@@ -21,6 +22,11 @@ client.renderCallback((result) => {
   console.log(`Page: ${page += 1}`)
   result.forEach((value) => {
     console.log(`${value.identifier}--${count += 1}`)
-    value.renderCallback(0, Infinity, (blob) => console.log(blob))
+    // value.renderCallback(0, Infinity, (blob) => console.log(blob))
   })
 }).finally(() => console.log('Done'))
+
+const result = new Result('nrps1974-04-13.late.aud.moore.berger.100289.flac16')
+result.renderCallback(0, Infinity, (blob) => {
+  console.log(blob)
+})
